Consolidate upload logic
