+++ pages/file_upload.py
@@ -26,7 +26,10 @@ from services.device_learning_service import (
     DeviceLearningService,
     get_device_learning_service,
 )
-from services.upload_service import process_uploaded_file, create_file_preview
+from services.data_processing.file_processor import (
+    process_uploaded_file,
+    create_file_preview,
+)
 from utils.upload_store import uploaded_data_store as _uploaded_data_store
 from services.upload_data_service import (
     get_uploaded_data as service_get_uploaded_data,
