+++ tests/test_process_and_analyze.py
@@ -3,7 +3,6 @@ import pandas as pd
 from services.data_processing.file_processor import FileProcessor
 
 from analytics.upload_processor import UploadAnalyticsProcessor
-from services.file_processing_service import FileProcessingService
 from services.data_validation import DataValidationService
 from services.data_processing.processor import Processor
 from services.data_processing.unified_file_validator import UnifiedFileValidator
@@ -14,8 +13,7 @@ def _create_components():
     vs = DataValidationService()
     dls = Processor(validator=vs)
     ufv = UnifiedFileValidator()
-    fps = FileProcessingService()
-    ua = UploadAnalyticsProcessor(fps, vs, dls, ufv)
+    ua = UploadAnalyticsProcessor(vs, dls)
     return fp, ua
